pr(Suspense): исправлен иморт lazy, Suspense вынесен в HomePage
